blogproject/backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
djoser==2.2.2
asgiref==3.7.2
sqlparse==0.4.4
pytz==2024.1
Pillow==10.2.0
djangorestframework-simplejwt==5.3.1
django-storages==1.14.2
boto3==1.34.49
django-allauth==0.61.1
django-cors-headers==4.3.1
django-allauth==0.61.1
PyJWT==2.8.0